# Improve TMap transformations