--- a/core/shared/src/main/scala/zio/stm/TArray.scala
+++ b/core/shared/src/main/scala/zio/stm/TArray.scala
@@ -255,6 +255,12 @@ final class TArray[A] private[stm] (private[stm] val array: Array[TRef[A]]) exte
     }
 
   /**
+   * Collects all elements into a list.
+   */
+  def toList: USTM[List[A]] =
+    STM.collectAll(array.map(_.get))
+
+  /**
    * Atomically updates all elements using a pure function.
    */
   def transform(f: A => A): USTM[Unit] =