--- a/core/shared/src/main/scala/zio/stm/TMap.scala
+++ b/core/shared/src/main/scala/zio/stm/TMap.scala
@@ -169,13 +169,12 @@ final class TMap[K, V] private (
    */
   def transform(f: (K, V) => (K, V)): USTM[Unit] =
     tBuckets.get.flatMap { buckets =>
-      buckets.toList.flatMap { list =>
+      buckets.toList.flatMap { data =>
         val g          = f.tupled
         val capacity   = buckets.array.length
-        val currData   = list.flatten
         val newBuckets = Array.fill[List[(K, V)]](capacity)(Nil)
 
-        val it = currData.iterator
+        val it = data.flatten.iterator
         while (it.hasNext) {
           val newPair = g(it.next)
           val idx     = TMap.indexOf(newPair._1, capacity)
@@ -201,10 +200,10 @@ final class TMap[K, V] private (
    */
   def transformM[E](f: (K, V) => STM[E, (K, V)]): STM[E, Unit] =
     tBuckets.get.flatMap { buckets =>
-      buckets.toList.flatMap { list =>
+      buckets.toList.flatMap { data =>
         val g = f.tupled
 
-        STM.foreach(list.flatten)(g).flatMap { mappedData =>
+        STM.foreach(data.flatten)(g).flatMap { mappedData =>
           val capacity   = buckets.array.length
           val newBuckets = Array.fill[List[(K, V)]](capacity)(Nil)